% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_section.R
\name{aggregate_META_information}
\alias{aggregate_META_information}
\title{Aggregate required information from VCF comment section}
\usage{
aggregate_META_information(comment_section)
}
\arguments{
\item{comment_section}{}
}
\description{
Aggregate required information from VCF comment section
}
