% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{apply_FUNC_functions}
\alias{apply_FUNC_functions}
\title{Apply functions to extract information from FUNC to vcf}
\usage{
apply_FUNC_functions(vavcf)
}
\arguments{
\item{vcffilepath}{}
}
\description{
Apply functions to extract information from FUNC to vcf
}
