% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_section.R
\name{attach_ID}
\alias{attach_ID}
\title{Attach sample ID from original vcf path and analysis name to output dataframes.
These entries serve as unique identifiers across tables}
\usage{
attach_ID(dataframe_oi, vcf_file, analysis_name)
}
\arguments{
\item{analysis_name}{}
}
\description{
Attach sample ID from original vcf path and analysis name to output dataframes.
These entries serve as unique identifiers across tables
}
