% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_VCF.R
\name{cdf_extract}
\alias{cdf_extract}
\title{Return dataframe from percentile and value pair in one string}
\usage{
cdf_extract(percentile_value)
}
\arguments{
\item{percentile_value}{}
}
\description{
Return dataframe from percentile and value pair in one string
}
