% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_based_filtering.R
\name{check_non_zero_AF}
\alias{check_non_zero_AF}
\title{Check AF entries in vcf INFO entries}
\usage{
check_non_zero_AF(vcfinfo)
}
\arguments{
\item{vcfinfo}{}
}
\description{
Check AF entries in vcf INFO entries
}
