% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_FUNC.R
\name{clean_FUNC_character}
\alias{clean_FUNC_character}
\title{Remove unwanted characters from vcf FUNC character vector}
\usage{
clean_FUNC_character(character_vector)
}
\arguments{
\item{character_vector}{}
}
\description{
Remove unwanted characters from vcf FUNC character vector
}
