% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_VCF.R
\name{cnv_extract_cdf}
\alias{cnv_extract_cdf}
\title{Parse CDF values from CDF_MAPD column to long and then wide data format.}
\usage{
cnv_extract_cdf(cnvrow_element)
}
\arguments{
\item{cnvrow_element}{}
}
\description{
Parse CDF values from CDF_MAPD column to long and then wide data format.
}
