% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_multi_entries.R
\name{combine_and_split}
\alias{combine_and_split}
\title{Collapses string and splits based on curly brackets in string.
Returns single list with splits as elements}
\usage{
combine_and_split(vcf_element_FUNC)
}
\arguments{
\item{vcf_element_FUNC}{}
}
\description{
Collapses string and splits based on curly brackets in string.
Returns single list with splits as elements
}
