% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_FUNC.R
\name{combine_FUNC_extracts}
\alias{combine_FUNC_extracts}
\title{Call all vcf FUNC extract functions}
\usage{
combine_FUNC_extracts(character_vector)
}
\arguments{
\item{character_vector}{}
}
\description{
Call all vcf FUNC extract functions
}
