% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_metrics_and_gene.R
\name{combine_function_calls}
\alias{combine_function_calls}
\title{Combines function call}
\usage{
combine_function_calls(vcfrow)
}
\arguments{
\item{vcfrow}{}
}
\description{
Combines function call
}
