% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_multi_entries.R, R/combine_results.R
\name{combine_orig_with_FUNC_extracts}
\alias{combine_orig_with_FUNC_extracts}
\title{Extract variant information and combine with read metrics for 'data-complete' table}
\usage{
combine_orig_with_FUNC_extracts(vcf_element)

combine_orig_with_FUNC_extracts(vcf_element)
}
\arguments{
\item{vcf_element}{}

\item{vcf}{}
}
\description{
Extract variant information and combine with read metrics for 'data-complete' table

Column bind vcf FUNC extracts with original vcf row
}
