% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_multi_entries.R
\name{entry_No_check}
\alias{entry_No_check}
\title{Check how many times 'gene' information is provided.
If more than once, 'combine_and_split' function is applied}
\usage{
entry_No_check(vcf_element)
}
\arguments{
\item{vcf_element}{}
}
\description{
Check how many times 'gene' information is provided.
If more than once, 'combine_and_split' function is applied
}
