% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_metrics_and_gene.R
\name{extract_format_values}
\alias{extract_format_values}
\title{Extract all values from FORMAT column}
\usage{
extract_format_values(vcf_row)
}
\arguments{
\item{vcf_row}{}
}
\description{
Extract all values from FORMAT column
}
