% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_FUNC.R
\name{extract_FUNC_coding}
\alias{extract_FUNC_coding}
\title{Extract coding information from FUNC vcf field}
\usage{
extract_FUNC_coding(character_vector)
}
\arguments{
\item{character_vector}{}
}
\description{
Extract coding information from FUNC vcf field
}
