% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_FUNC.R
\name{extract_FUNC_exon}
\alias{extract_FUNC_exon}
\title{Extract location information from FUNC vcf field}
\usage{
extract_FUNC_exon(character_vector)
}
\arguments{
\item{character_vector}{}
}
\description{
Extract location information from FUNC vcf field
}
