% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_FUNC.R
\name{extract_FUNC_gene}
\alias{extract_FUNC_gene}
\title{Extract Gene symbol from FUNC vcf field}
\usage{
extract_FUNC_gene(character_vector)
}
\arguments{
\item{character_vector}{}
}
\description{
Extract Gene symbol from FUNC vcf field
}
