% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_FUNC.R
\name{extract_FUNC_origAlt}
\alias{extract_FUNC_origAlt}
\title{Extract origAlt information from FUNC vcf field}
\usage{
extract_FUNC_origAlt(character_vector)
}
\arguments{
\item{character_vector}{}
}
\description{
Extract origAlt information from FUNC vcf field
}
