% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_FUNC.R
\name{extract_FUNC_origPos}
\alias{extract_FUNC_origPos}
\title{Extract origPos information from FUNC vcf field}
\usage{
extract_FUNC_origPos(character_vector)
}
\arguments{
\item{character_vector}{}
}
\description{
Extract origPos information from FUNC vcf field
}
