% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_FUNC.R
\name{extract_FUNC_protein}
\alias{extract_FUNC_protein}
\title{Extract protein information from FUNC vcf field}
\usage{
extract_FUNC_protein(character_vector)
}
\arguments{
\item{character_vector}{}
}
\description{
Extract protein information from FUNC vcf field
}
