% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_FUNC.R
\name{extract_FUNC_transcript}
\alias{extract_FUNC_transcript}
\title{Extract transcript information from FUNC vcf field}
\usage{
extract_FUNC_transcript(character_vector)
}
\arguments{
\item{character_vector}{}
}
\description{
Extract transcript information from FUNC vcf field
}
