% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFO_extract.R
\name{extract_gene_parameters}
\alias{extract_gene_parameters}
\title{Extract values from FUNC string}
\usage{
extract_gene_parameters(gene_string)
}
\arguments{
\item{info_string}{}
}
\description{
Extract values from FUNC string
}
