% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variant_info.R
\name{extract_variant_info}
\alias{extract_variant_info}
\title{Extra#'#'#'racExtract character string from INFO field in VCF}
\usage{
extract_variant_info(vcf_info)
}
\arguments{
\item{vcf_info}{}
}
\description{
Extra#'#'#'racExtract character string from INFO field in VCF
}
