% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{from_multiple_FUNC_entries_to_multiple_vcf_rows}
\alias{from_multiple_FUNC_entries_to_multiple_vcf_rows}
\title{Parse multiple entries from FUNC field and convert into multiple rows.
Following functions parse FUNC field then uniformly.w}
\usage{
from_multiple_FUNC_entries_to_multiple_vcf_rows(vcf_element)
}
\arguments{
\item{vcf_element}{}
}
\description{
Parse multiple entries from FUNC field and convert into multiple rows.
Following functions parse FUNC field then uniformly.w
}
