% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_FUNC.R
\name{FUNC_extracts_to_df}
\alias{FUNC_extracts_to_df}
\title{Call all vcf FUNC extract functions}
\usage{
FUNC_extracts_to_df(character_vector)
}
\arguments{
\item{character_vector}{}
}
\description{
Call all vcf FUNC extract functions
}
