% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFO_gene_extract.R
\name{gene_extract}
\alias{gene_extract}
\title{Extract all shortest possible strings within curly brackets}
\usage{
gene_extract(info_string)
}
\arguments{
\item{func_string}{}
}
\description{
Extract all shortest possible strings within curly brackets
}
\examples{
{gene:string}
}
