% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath.R
\name{generate_paths}
\alias{generate_paths}
\title{Generate paths for output files based on vcfpath}
\usage{
generate_paths(vcfpath)
}
\arguments{
\item{vcfpath}{}
}
\description{
Generate paths for output files based on vcfpath
}
