% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tables.R
\name{interpretation_table}
\alias{interpretation_table}
\title{Output variant table to be used for variant interpretation
TSGinfo, cancerHotspots etc.}
\usage{
interpretation_table(combined_output)
}
\arguments{
\item{combined_output}{}
}
\description{
Output variant table to be used for variant interpretation
TSGinfo, cancerHotspots etc.
}
