% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tables.R
\name{metric_table}
\alias{metric_table}
\title{Output clean metric table starting from combined output from VCF parsing}
\usage{
metric_table(combined_output)
}
\arguments{
\item{combined_output}{}
}
\description{
Output clean metric table starting from combined output from VCF parsing
}
