% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_VCF.R
\name{parse_VCF}
\alias{parse_VCF}
\title{Calls Annotation Module functions for cancerHotspots, COSMIC, GNOMAD, ClinVar}
\usage{
parse_VCF(vcf, annotation_fp)
}
\arguments{
\item{annotation_fp}{}
}
\description{
Calls Annotation Module functions for cancerHotspots, COSMIC, GNOMAD, ClinVar
}
