% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_VCF.R
\name{parse_vcfpath_return_metainformation}
\alias{parse_vcfpath_return_metainformation}
\title{Parse VCF -- create CNV}
\usage{
parse_vcfpath_return_metainformation(vcfpath)
}
\arguments{
\item{vcf}{}
}
\description{
Parse VCF -- create CNV
}
