% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsed_filepaths.R
\name{parsed_filepaths}
\alias{parsed_filepaths}
\title{Generate paths for output files based on vcfpath}
\usage{
parsed_filepaths(output_path)
}
\arguments{
\item{vcfpath}{}
}
\description{
Generate paths for output files based on vcfpath
}
