% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF.R
\name{read_vcf}
\alias{read_vcf}
\title{Read VCF file skipping ## commented entries.
Also skips entries with 'AF=0;' as these are likely only derived by hotspot annotation and not by variant calling.}
\usage{
read_vcf(vcf_filepath)
}
\arguments{
\item{vcf_filepath}{}
}
\description{
Read VCF file skipping ## commented entries.
Also skips entries with 'AF=0;' as these are likely only derived by hotspot annotation and not by variant calling.
}
