% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF.R
\name{readVCF}
\alias{readVCF}
\title{Uses VariantAnnotation package to read in VCF as GRanges object}
\usage{
readVCF(vcffilepath)
}
\arguments{
\item{vcffilepath}{}
}
\description{
Uses VariantAnnotation package to read in VCF as GRanges object
}
