% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_leftover_chars.R
\name{remove_multientry_leftover_characters}
\alias{remove_multientry_leftover_characters}
\title{Introduce function to remove buggy output}
\usage{
remove_multientry_leftover_characters(tableofinterest)
}
\arguments{
\item{tableofinterest}{}
}
\description{
Introduce function to remove buggy output
}
