% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_based_filtering.R
\name{return_filtered_vectors}
\alias{return_filtered_vectors}
\title{Extract comma separated values with specified indeces}
\usage{
return_filtered_vectors(char_string, vector_indeces)
}
\arguments{
\item{vector_indeces}{}
}
\description{
Extract comma separated values with specified indeces
}
