% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_metrics_and_gene.R
\name{standard_columns}
\alias{standard_columns}
\title{Extract VCF standard columns. Additionally extract sample name from column 9 and remove column}
\usage{
standard_columns(vcf)
}
\arguments{
\item{vcf}{}
}
\description{
Extract VCF standard columns. Additionally extract sample name from column 9 and remove column
}
