% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variant_info.R
\name{variant_info_to_row}
\alias{variant_info_to_row}
\title{Convert character string to single row with multiple columns per variant info string}
\usage{
variant_info_to_row(vcf_info_string)
}
\arguments{
\item{vcf_info_string}{}
}
\description{
Convert character string to single row with multiple columns per variant info string
}
