% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tables.R
\name{variant_table}
\alias{variant_table}
\title{Output clean variant table starting from combined output from VCF parsing}
\usage{
variant_table(combined_output)
}
\arguments{
\item{combined_output}{}
}
\description{
Output clean variant table starting from combined output from VCF parsing
}
