% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_section.R
\name{vcf_comment_section}
\alias{vcf_comment_section}
\title{Extract commented section from VCF}
\usage{
vcf_comment_section(vcfpath)
}
\arguments{
\item{vcfpath}{}
}
\description{
Extract commented section from VCF
}
