% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_section.R
\name{write_out_META_information}
\alias{write_out_META_information}
\title{WRITE OUT meta information from VCF comment section to yaml}
\usage{
write_out_META_information(meta_information, basedir)
}
\arguments{
\item{basedir}{}
}
\description{
WRITE OUT meta information from VCF comment section to yaml
}
