% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_tables.R
\name{write_output_tables}
\alias{write_output_tables}
\title{Creates output tables and directory by individual function calls}
\usage{
write_output_tables(combined_output, filepath)
}
\arguments{
\item{filepath}{}
}
\description{
Creates output tables and directory by individual function calls
}
