% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_VCF.R
\name{write_parsed_fp_txt}
\alias{write_parsed_fp_txt}
\title{Write out parsed filepath information}
\usage{
write_parsed_fp_txt(parsed_fp, vcf_file, analysis_name)
}
\arguments{
\item{parsed_fp}{}
}
\description{
Write out parsed filepath information
}
