% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_parsed_output.R
\name{write_parsed_output}
\alias{write_parsed_output}
\title{Write out VCF parse information}
\usage{
write_parsed_output(vcf, vcf_path, parsed_fp)
}
\arguments{
\item{parsed_fp}{}
}
\description{
Write out VCF parse information
}
