% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC_multi_entries.R
\name{combine_orig_with_FUNC_extracts}
\alias{combine_orig_with_FUNC_extracts}
\title{Extract variant information and combine with read metrics for 'data-complete' table}
\usage{
combine_orig_with_FUNC_extracts(vcf)
}
\arguments{
\item{vcf}{}
}
\description{
Extract variant information and combine with read metrics for 'data-complete' table
}
