% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_results.R
\name{combine_orig_with_FUNC_extracts}
\alias{combine_orig_with_FUNC_extracts}
\title{Column bind vcf FUNC extracts with original vcf row}
\usage{
combine_orig_with_FUNC_extracts(vcf_element)
}
\arguments{
\item{vcf_element}{}
}
\description{
Column bind vcf FUNC extracts with original vcf row
}
